fix crash if no synaptic config is there (LP: #624703)

+++ GDebi/GDebi.py
@@ -607,6 +607,8 @@ Install software from trustworthy software distributors only.
         return False
         
     def write_synaptic_config_file(self, config, path):
+        if not os.path.exists(config):
+            return
         config_file = open(path, "w")
         config_file.write(config.dump())
         config_file.close()
